// FrontEnd/ClientApp/src/app/models/ICompany.ts
export interface ICompany {
  idCompany: number;
  identificationNumber: number;
  nitCompany: number;
  companyName: string;
  firtsName: string;
  secondName: string;
  firtsLastName: string;
  secondLastName: string;
  email: string;
  phone: string;
  authorizedSendEmail: boolean;
  authorizedSendPhone: boolean;
  idIdentificationType: number;
}